more refactoring to separate tests
