+++ tests/functional/Tribe/Events/Importer/File_Importer_EventsTest.php
@@ -8,6 +8,11 @@ use Tribe__Events__Importer__File_Importer_Events as Events_Importer;
 
 class File_Importer_EventsTest extends \Codeception\TestCase\WPTestCase {
 
+	/**
+	 * @var \Tribe__Events__Importer__Featured_Image_Uploader
+	 */
+	protected $featured_image_uploader;
+
 	protected $field_map = [
 		'event_name',
 		'event_description',
@@ -50,11 +55,6 @@ class File_Importer_EventsTest extends \Codeception\TestCase\WPTestCase {
 	 */
 	protected $template = 'events';
 
-	/**
-	 * @var \Tribe__Events__Importer__Featured_Image_Uploader
-	 */
-	protected $featured_image_uploader;
-
 	public function setUp() {
 		// before
 		parent::setUp();
@@ -77,49 +77,6 @@ class File_Importer_EventsTest extends \Codeception\TestCase\WPTestCase {
 	 */
 	public function it_should_be_instantiatable() {
 		$sut = $this->make_instance();
-	}
-
-	/**
-	 * @test
-	 * it should not mark record as invalid if featured image entry is missing
-	 */
-	public function it_should_not_mark_record_as_invalid_if_featured_image_entry_is_missing() {
-		$sut = $this->make_instance( 'featured-image' );
-
-		$post_id = $sut->import_next_row();
-
-		$this->assertNotFalse( $post_id );
-	}
-
-	/**
-	 * @test
-	 * it should import and attach featured image if featured image is ok
-	 */
-	public function it_should_import_and_attach_featured_image_if_featured_image_is_ok() {
-		$image_url     = plugins_url( '_data/csv-import-test-files/featured-image/images/featured-image.jpg', codecept_data_dir() );
-		$attachment_id = $this->factory()->attachment->create_upload_object( $image_url );
-		$this->featured_image_uploader->upload_and_get_attachment()->willReturn( $attachment_id );
-
-		$sut = $this->make_instance( 'featured-image' );
-
-		$post_id = $sut->import_next_row();
-
-		$this->assertEquals( $attachment_id, get_post_thumbnail_id( $post_id ) );
-	}
-
-	/**
-	 * @test
-	 * it should not import and attach featured image if featured image is not ok
-	 */
-	public function it_should_not_import_and_attach_featured_image_if_featured_image_is_not_ok() {
-		$this->featured_image_uploader->upload_and_get_attachment()->willReturn( false );
-
-		$sut = $this->make_instance( 'featured-image' );
-
-		$post_id = $sut->import_next_row();
-
-		$has_thumbnail = wp_get_attachment_url( get_post_thumbnail_id( $post_id ) );
-		$this->assertFalse( $has_thumbnail );
 	}
 
 	protected function make_instance( $template_dir = null ) {
